ceph: allow to deploy python-cradox
python-cradox is a new dependency in Gnocchi when deploying Ceph
backend.
RDO team does not want to handle this dependency because it's too
specific to the Ceph backend, so puppet-gnocchi will be able to deploy
it or not (installed by default).

Usage of puppetlabs-stdlib will allow duplicated resource, also using
the tags will allow to modify the resource.
Adding tests & release note.

Change-Id: I110d1080b2dcf362f4e48a773ceef1e87d68b7cd